common/shstr.c, common/arch.c, common/path.c: Fix some int that should have been   size_t, found using splint. common/arch.c: Also fix an "int" that should have been "unsigned long". include/libproto.h, include/path.h: Update prototypes where needed for these int   to size_t changes.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11278 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/arch.c
@@ -173,7 +173,7 @@ archetype *get_archetype_by_type_subtype(int type, int subtype) {
 object *create_archetype_by_object_name(const char *name) {
     archetype *at;
     char tmpname[MAX_BUF];
-    int i;
+    size_t i;
 
     strncpy(tmpname, name, MAX_BUF-1);
     tmpname[MAX_BUF-1] = 0;
@@ -714,7 +714,7 @@ archetype *find_archetype(const char *name) {
  * Will call fatal() if archetype table is too small to contain archetypes.
  */
 static void add_arch(archetype *at) {
-    int index = hasharch(at->name, ARCHTABLE), org_index = index;
+    unsigned long index = hasharch(at->name, ARCHTABLE), org_index = index;
 
     for (;;) {
         if (arch_table[index] == NULL) {
